requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bcrypt==3.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-core==0.2.36
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.107
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.23.0
websockets==13.0
yarl==1.9.4
